# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.CreateUserTokenPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  main.ForgotPasswordPayload:
    properties:
      email:
        maxLength: 255
        type: string
    required:
    - email
    type: object
  main.RegisterUserPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 3
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
  main.ResetPasswordPayload:
    properties:
      password:
        maxLength: 72
        minLength: 3
        type: string
    required:
    - password
    type: object
...
  main.UserWithToken:
    properties:
      coin:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      token:
        type: string
      token_version:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  swagger.EnvelopeError:
    properties:
      error:
        type: string
    type: object
  swagger.EnvelopeString:
    properties:
      data:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for an online novel platform — read, publish, and manage
    digital stories.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Govel API
paths:
  /authentication/forgot-password:
    post:
      consumes:
      - application/json
      description: Send reset password link to user's email
      parameters:
      - description: Email payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ForgotPasswordPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Plain reset token
          schema:
            $ref: '#/definitions/swagger.EnvelopeString'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: user npt found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      summary: Request forgot password
      tags:
      - authentication
  /authentication/reset-password/{token}:
    patch:
      consumes:
      - application/json
      description: Reset password using token from forgot-password email
      parameters:
      - description: Reset password token
        in: path
        name: token
        required: true
        type: string
      - description: New password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ResetPasswordPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed message
          schema:
            $ref: '#/definitions/swagger.EnvelopeString'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: invalid token
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      summary: Reset user password
      tags:
      - authentication
  /authentication/token:
    post:
      consumes:
      - application/json
      description: Creates a token for a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserTokenPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Token
          schema:
            $ref: '#/definitions/swagger.EnvelopeString'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      summary: Creates a token
      tags:
      - authentication
  /authentication/user:
    post:
      consumes:
      - application/json
      description: Registers a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: user registered
          schema:
            $ref: '#/definitions/main.UserWithToken'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      summary: Registers a user
      tags:
      - authentication
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"